requirements.txt
django 
django_taggit
Markdown
psycopg2
psycopg2-binary
python-dotenv
